이전 커밋에서 변경된 내용에따라 연관된 코드 변경
- useGetProductList... 커스텀 훅을 useGetCategoryInfo로 변경 및 인자값 변경
- 관련 import 링크 및 모듈 변수이름 변경.
- 기타 등등변경, 아직 실행 및 디버깅은 안해봄.
