+++ react-front-app/src/store.js
@@ -1,10 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { categorySlice } from './services/categorySlice';
-import { productListByISlice } from './services/productListByISlice';
+import { categoryInfoSlice } from './services/categoryInfoSlice';
 
 export const store = configureStore({
   reducer: {
     category: categorySlice.reducer,
-    productListByI : productListByISlice.reducer
+    categoryInfo : categoryInfoSlice.reducer
   },
 });
